+++ src/main/java/com/cuadratura/app/service/impl/CargaWmsServiceImpl.java
@@ -23,7 +23,7 @@ private static final Logger LOGGER = LogManager.getLogger(CargaWmsServiceImpl.cl
 		return cargaWmsRepository;
 	}
 	
-	public Integer saveCargaWms(CargaWms cargaWms) {
+	public Long saveCargaWms(CargaWms cargaWms) {
 		LOGGER.info("insert");
 		return cargaWmsRepository.saveCargaWms(cargaWms);
 	}
